codex(refactor): extract compose render context
